requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
opencv-python>=4.5.0

# Data processing and visualization
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Configuration and utilities
PyYAML>=5.4.0
tqdm>=4.62.0
tensorboard>=2.7.0

# Development and testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.7.0
flake8>=3.9.0
isort>=5.9.0

# Optional: for advanced features
albumentations>=1.1.0
timm>=0.4.12
wandb>=0.12.0

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0